Old frontend deleted. Back-end completed.
