--- frontend/src/utils/dateUtils.ts
@@ -1,18 +1,0 @@
-export const formatDueDate = (dueDate: string): string => {
-    const today = new Date();
-    const due = new Date(dueDate);
-    const diffTime = due.getTime() - today.getTime();
-    const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24));
-  
-    if (diffDays === 0) {
-      return "Today";
-    } else if (diffDays === 1) {
-      return "Tomorrow";
-    } else if (diffDays === -1) {
-      return "Yesterday";
-    } else if (diffDays < 0) {
-      return `${Math.abs(diffDays)} days ago`;
-    } else {
-      return `In ${diffDays} days`;
-    }
-  };
