--- frontend/types.ts
@@ -1,7 +1,0 @@
-export interface Todo {
-    id: string;
-    title: string;
-    description: string;
-    dueDate: string; //YYYY-MM-DD
-    completed: boolean;
-  }
